LPS-105024 Move builders to one file
