--- modules/apps/portal-crypto/portal-crypto-hash-api/src/main/java/com/liferay/portal/crypto/hash/request/builder/PepperBuilder.java
@@ -1,26 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.crypto.hash.request.builder;
-
-import com.liferay.portal.crypto.hash.request.command.pepper.PepperCommand;
-
-/**
- * @author Carlos Sierra Andrés
- */
-public interface PepperBuilder extends SaltBuilder {
-
-	public SaltBuilder pepperCommand(PepperCommand pepperCommand);
-
-}
