--- modules/apps/portal-crypto/portal-crypto-hash-api/src/main/resources/com/liferay/portal/crypto/hash/request/builder/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
